pyproject.toml
[project]
name = "pdf-obfuscator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Antoine LEMAITRE", email = "antoine.lemaitre@jobteaser.com" }
]
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.39.4",
    "fastapi>=0.116.0",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.3",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
    "numpy>=2.3.1",
    "pypdfium2>=4.30.0",
    "Pillow>=10.0.0",
    "reportlab>=4.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
]
